@@ -15,28 +15,28 @@
         "ext-json": "*",
         "doctrine/dbal": "^2.9",
         "dompdf/dompdf": "^0.8",
-        "fideloper/proxy": "^4.2",
-        "graham-campbell/markdown": "11.1.0",
-        "guzzlehttp/guzzle": "^6.4",
+        "graham-campbell/markdown": "*",
+        "guzzlehttp/guzzle": "^7.0",
         "guzzlehttp/psr7": "^1.6",
+        "http-interop/http-factory-guzzle": "^1.2",
         "intervention/image": "^2.5",
         "laracasts/utilities": "^3.0",
-        "laravel/framework": "^6.5",
-        "laravel/socialite": "^4.3",
-        "laravel/tinker": "^1.0",
+        "laravel/framework": "^10.0",
+        "laravel/socialite": "^5.0",
+        "laravel/tinker": "^2.0",
         "laravelcollective/html": "^6.0",
-        "league/flysystem-aws-s3-v3": "^1.0",
-        "league/omnipay": "^3.0",
+        "league/flysystem-aws-s3-v3": "^3.0",
+        "league/omnipay": "^3.2",
         "maatwebsite/excel": "^3.1",
-        "mcamara/laravel-localization": "^1.4",
+        "mcamara/laravel-localization": "^2.0",
         "mews/purifier": "^3.1",
-        "milon/barcode": "^6.0",
+        "milon/barcode": "*",
         "nitmedia/wkhtml2pdf": "dev-master",
         "omnipay/common": "^3.0",
         "omnipay/dummy": "^3.0",
         "omnipay/paypal": "^3.0",
         "omnipay/stripe": "^3.1",
-        "php-http/curl-client": "^1.7",
+        "php-http/curl-client": "^2.2.1",
         "php-http/message": "^1.6",
         "predis/predis": "~1.1",
         "stripe/stripe-php": "^7.1",
@@ -45,12 +45,12 @@
     },
     "require-dev": {
         "barryvdh/laravel-ide-helper": "^2.6",
-        "facade/ignition": "^1.4",
-        "fzaninotto/faker": "1.9.*",
+        "fakerphp/faker": "^1.20",
         "mockery/mockery": "^1.0",
-        "nunomaduro/collision": "^3.0",
-        "phpspec/phpspec": "^6.0",
-        "phpunit/phpunit": "^8.0",
+        "nunomaduro/collision": "*",
+        "phpspec/phpspec": "^7.0",
+        "phpunit/phpunit": "^9.0",
+        "spatie/laravel-ignition": "*",
         "symfony/css-selector": "~3.0",
         "symfony/dom-crawler": "~3.0"
     },
@@ -102,7 +102,10 @@
     "config": {
         "optimize-autoloader": true,
         "preferred-install": "dist",
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "php-http/discovery": true
+        }
     },
     "extra": {
         "laravel": {
