@@ -2,24 +2,25 @@
 
 namespace App\Http\Middleware;
 
+use Illuminate\Http\Middleware\TrustProxies as Middleware;
 use Illuminate\Http\Request;
-use Fideloper\Proxy\TrustProxies as Middleware;
 
 class TrustProxies extends Middleware
 {
     /**
      * The trusted proxies for this application.
      *
-     * @var array|string
+     * @var string|array
      */
     protected $proxies = [
-      '192.168.1.1',
-      '192.168.1.2',
-	];
+        '192.168.1.1',
+        '192.168.1.2',
+    ];
+
     /**
      * The headers that should be used to detect proxies.
      *
-     * @var string
+     * @var int
      */
-    protected $headers = Request::HEADER_X_FORWARDED_ALL;
+    protected $headers = Request::HEADER_X_FORWARDED_FOR | Request::HEADER_X_FORWARDED_HOST | Request::HEADER_X_FORWARDED_PORT | Request::HEADER_X_FORWARDED_PROTO;
 }
