@@ -2,7 +2,7 @@
 
 return [
 
-    'debug'       => env('APP_DEBUG_PDF', false),
+    'debug'       => env('APP_DEBUG_PDF', true),
     'binpath'     => 'lib/',
     'binfile'     => env('WKHTML2PDF_BIN_FILE', 'wkhtmltopdf-amd64'),
     'output_mode' => 'I',
